+++ uamqp/async_ops/mgmt_operation_async.py
@@ -6,7 +6,6 @@
 
 import logging
 import asyncio
-import functools
 import uuid
 
 #from uamqp.session import Session
@@ -103,4 +102,4 @@ class MgmtOperationAsync(MgmtOperation):
 
     async def destroy_async(self):
         """Close the send/receive links for this node asynchronously."""
-        await self.loop.run_in_executor(None, functools.partial(self._mgmt_op.destroy))
+        self._mgmt_op.destroy()