+++ uamqp/async_ops/sender_async.py
@@ -6,7 +6,6 @@
 
 import asyncio
 import logging
-import functools
 
 from uamqp import sender
 from uamqp import errors, constants
@@ -112,7 +111,7 @@ class MessageSenderAsync(sender.MessageSender):
 
     async def destroy_async(self):
         """Asynchronously close both the Sender and the Link. Clean up any C objects."""
-        await self.loop.run_in_executor(None, functools.partial(self.destroy))
+        self.destroy()
 
     async def open_async(self):
         """Asynchronously open the MessageSender in order to start
@@ -123,7 +122,7 @@ class MessageSenderAsync(sender.MessageSender):
          or the credentials are rejected.
         """
         try:
-            await self.loop.run_in_executor(None, functools.partial(self._sender.open))
+            self._sender.open()
         except ValueError:
             raise errors.AMQPConnectionError(
                 "Failed to open Message Sender. "
@@ -162,4 +161,4 @@ class MessageSenderAsync(sender.MessageSender):
 
     async def close_async(self):
         """Close the sender asynchronously, leaving the link intact."""
-        await self.loop.run_in_executor(None, functools.partial(self._sender.close))
+        self._sender.close()