+++ uamqp/async_ops/session_async.py
@@ -6,7 +6,6 @@
 
 import asyncio
 import logging
-import functools
 
 from uamqp import session
 from uamqp import constants
@@ -116,4 +115,4 @@ class SessionAsync(session.Session):
         """
         for _, link in self._mgmt_links.items():
             await link.destroy_async()
-        await self.loop.run_in_executor(None, functools.partial(self._session.destroy))
+        self._session.destroy()