+++ uamqp/connection.py
@@ -184,7 +184,7 @@ class Connection:
                 description = b"Connection in an unexpected error state."
                 self._error = errors._process_connection_error(self.error_policy, condition, description, None)  # pylint: disable=protected-access
 
-    def lock(self, timeout=10.0):
+    def lock(self, timeout=3.0):
         if not self._lock.acquire(timeout=timeout):  # pylint: disable=unexpected-keyword-arg
             raise TimeoutError("Failed to acquire connection lock.")
 
@@ -193,8 +193,12 @@ class Connection:
             self._lock.release()
         except RuntimeError:
             pass
-        except KeyboardInterrupt:
-            self.release()
+        except:
+            _logger.debug("Got error when attempting to release connection lock.")
+            try:
+                self._lock.release()
+            except RuntimeError:
+                pass
             raise
 
     def destroy(self):